+++ src/WebAudio/Player.js
@@ -21,11 +21,24 @@ export default class Player extends AudioContext {
   audioSource = null;
 
   /**
+   * Private object to hold GainNode.
+   * @private
+   * @type {GainNode}
+   */
+  gainNode = null;
+
+  /**
    * Player playback state.
    * @private
    * @type {boolean}
    */
   playing = false;
+
+  /**
+   * Player volume. 0 is muted, 1 is 100%.
+   * @type {number}
+   */
+  volume = 1;
 
   /**
    * Point in time where player is paused.
@@ -42,6 +55,18 @@ export default class Player extends AudioContext {
   startedAt = 0;
 
   /**
+   * Player constructor.
+   * Creates a GainNode and stores it.
+   */
+  constructor() {
+    super();
+
+    this.gainNode = this.audioContext.createGain();
+    this.gainNode.gain.value = this.volume;
+    this.gainNode.connect(this.audioContext.destination);
+  }
+
+  /**
    * Create and initialize the AudioBufferSourceNode object.
    *
    */
@@ -54,11 +79,14 @@ export default class Player extends AudioContext {
     // Select what to play
     this.audioSource.buffer = this.audioBuffer;
 
-    // Connect to the speakers!
-    this.audioSource.connect(this.audioContext.destination);
+    // Connect to the GainNode!
+    this.audioSource.connect(this.gainNode);
 
     // Add some event handlers;
-    this.audioSource.addEventListener('ended', this.suspendAudioContext);
+    this.audioSource.addEventListener('ended', () => {
+      this.fireEvent('ended');
+      this.suspendAudioContext();
+    });
   }
 
   /**
@@ -68,7 +96,7 @@ export default class Player extends AudioContext {
   disconnectBufferSource() {
     if (this.audioSourceExists()) {
       this.audioSource.disconnect();
-      this.audioSource.removeEventListener('ended', this.suspendAudioContext);
+      this.audioSource = null;
     }
   }
 
@@ -97,6 +125,7 @@ export default class Player extends AudioContext {
       .then(response => response.arrayBuffer())
       .then(audioData => audioContext.decodeAudioData(audioData, decodedAudio => {
         this.audioBuffer = decodedAudio;
+        this.createBufferSource();
         this.fireEvent('loaded');
       }))
       .catch(error => {
@@ -128,7 +157,9 @@ export default class Player extends AudioContext {
   play() {
     const offset = this.pausedAt;
 
-    this.createBufferSource();
+    if (offset !== 0) {
+      this.createBufferSource();
+    }
 
     if (this.audioContext.state === 'suspended') {
       this.audioContext.resume();
@@ -195,6 +226,25 @@ export default class Player extends AudioContext {
   }
 
   /**
+   * Set the volume of the payer to a value between 0 - 1.
+   * 0 means no volume (muted), 1 means max.
+   *
+   * @param {number} volume - Value for volume between 0 an 1.
+   */
+  setVolume(volume = 1) {
+    this.volume = volume;
+    this.gainNode.gain.setValueAtTime(volume, this.audioContext.currentTime);
+  }
+
+  /**
+   * Mute the player by setting its volume to 0.
+   */
+  mute() {
+    this.volume = 0;
+    this.gainNode.gain.setValueAtTime(0, this.audioContext.currentTime);
+  }
+
+  /**
    * Get the currentTime for the audio that is loaded.
    *
    * @returns {number} - The currentTime value.
@@ -209,5 +259,24 @@ export default class Player extends AudioContext {
     }
 
     return 0;
+  }
+
+  /**
+   * Return the AudioBufferSourceNode node.
+   * Note that after a pause/resume the AudioBufferSourceNode will be recreated.
+   *
+   * @returns {AudioBufferSourceNode} - The current available AudioBufferSourceNode.
+   */
+  getBufferSource() {
+    return this.audioSource;
+  }
+
+  /**
+   * Return the AudioContext node.
+   *
+   * @returns {AudioContext} - The current available AudioContext.
+   */
+  getAudioContext() {
+    return this.audioContext;
   }
 }